Changed to chart.js library
