+++ water/package.json
@@ -19,10 +19,12 @@
     "@vx/shape": "0.0.197",
     "@vx/tooltip": "0.0.197",
     "apexcharts": "^3.19.2",
+    "chart.js": "^2.9.3",
     "d3-array": "^2.4.0",
     "next": "^9.4.4",
     "react": "^16.13.1",
     "react-apexcharts": "^1.3.7",
+    "react-chartjs-2": "^2.9.0",
     "react-dom": "^16.13.1",
     "react-scripts": "3.4.1",
     "socket.io-client": "^2.3.0"
